==> notebooks/Run2/extra/2022/v13/05_jodafons_emulate_electron_sequence.ipynb <==
# Kepler Framework, Emulate v13

from kepler.pandas.menu       import ElectronSequence as Chain
from kepler.pandas.readers    import load, load_in_loop, load_hdf
from core import Efficiency
from Gaugi import mkdir_p, progressbar
import numpy as np
import pandas as pd
import collections
import os
from pprint import pprint
from copy import deepcopy

import gc
%config InlineBackend.figure_format = 'retina'
%load_ext autoreload
%autoreload 2

output_path = 'output/emulation'
mkdir_p(output_path)

dpath_data17 = '../../data/output/data17_13TeV.AllPeriods.sgn.probes_lhvloose_EGAM1.bkg.vprobes_vlhvloose_EGAM7.GRL_v97.dataframe.h5'
dpath_data18 = '../../data/output/data18_13TeV.AllPeriods.sgn.probes_lhvloose_EGAM1.bkg.vprobes_vlhvloose_EGAM7.GRL_v97.dataframe.h5'
dpath_mc16   = '../../data/output/mc16_13TeV.302236_309995_341330.sgn.boosted_probes.WZ_llqq_plus_radion_ZZ_llqq_plus_ggH3000.merge.dataframe.h5'

tpath_data17 = 'output/emulation/data17_table_v13.h5'
tpath_data18 = 'output/emulation/data18_table_v13.h5'
tpath_mc16   = 'output/emulation/mc16_table_v13.h5'

## Setup Chains:

# create my chain
chains = [
            Chain( "HLT_e17_lhvloose_nod0_ringer_v13_L1EM15VHI", L1Seed = 'L1_EM15VHI' , l2calo_column = 'ringer_v13_vloose'),
            Chain( "HLT_e28_lhtight_nod0_ringer_v13_ivarloose", L1Seed = 'L1_EM22VHI' , l2calo_column = 'ringer_v13_tight' ),
            Chain( "HLT_e60_lhmedium_nod0_ringer_v13_L1EM24VHI", L1Seed = 'L1_EM24VHI' , l2calo_column = 'ringer_v13_medium'),
            Chain( "HLT_e140_lhloose_nod0_ringer_v13" , L1Seed = 'L1_EM24VHI', l2calo_column = 'ringer_v13_loose'),
]

def emulate( df ):
    for chain in progressbar(chains, prefix='Emulate...'):
        chain.apply(df)

## Emulate 2017:

Take something like 1 hour

data17_df = pd.concat( (load_hdf(dpath_data17), load_hdf(tpath_data17)), axis=1)

emulate(data17_df)

data17_df.shape

### Validate:

eff_data17_Zee = Efficiency( output_path+'/data17_Zee_efficiency_v13.root')

eff_data17_Zee.fill( data17_df.loc[data17_df['target']==1], 'HLT_e17_lhvloose_nod0_ringer_v13_L1EM15VHI', pidname='el_lhvloose', reset=True)
eff_data17_Zee.fill( data17_df.loc[data17_df['target']==1], 'HLT_e28_lhtight_nod0_ringer_v13_ivarloose' , pidname='el_lhtight' , reset=True)
eff_data17_Zee.fill( data17_df.loc[data17_df['target']==1], 'HLT_e60_lhmedium_nod0_ringer_v13_L1EM24VHI', pidname='el_lhmedium', reset=True)
eff_data17_Zee.fill( data17_df.loc[data17_df['target']==1], 'HLT_e140_lhloose_nod0_ringer_v13'          , pidname='el_lhloose' , reset=True)

eff_data17_Zee.save()

eff_data17_JF17 = Efficiency( output_path+'/data17_JF17_efficiency_v13.root')

eff_data17_JF17.fill( data17_df.loc[data17_df['target']==0], 'HLT_e17_lhvloose_nod0_ringer_v13_L1EM15VHI', pidname='!el_lhvloose',reset=True)
eff_data17_JF17.fill( data17_df.loc[data17_df['target']==0], 'HLT_e28_lhtight_nod0_ringer_v13_ivarloose' , pidname='!el_lhvloose',reset=True)
eff_data17_JF17.fill( data17_df.loc[data17_df['target']==0], 'HLT_e60_lhmedium_nod0_ringer_v13_L1EM24VHI', pidname='!el_lhvloose',reset=True)
eff_data17_JF17.fill( data17_df.loc[data17_df['target']==0], 'HLT_e140_lhloose_nod0_ringer_v13'          , pidname='!el_lhvloose',reset=True)

eff_data17_JF17.save()

del data17_df

## Emulate 2018:

data18_df = pd.concat( (load_hdf(dpath_data18), load_hdf(tpath_data18)), axis=1)

emulate(data18_df)

data18_df.shape

eff_data18_Zee = Efficiency( output_path+'/data18_Zee_efficiency_v13.root')

eff_data18_Zee.fill( data18_df.loc[data18_df['target']==1], 'HLT_e17_lhvloose_nod0_ringer_v13_L1EM15VHI', pidname='el_lhvloose', reset=True)
eff_data18_Zee.fill( data18_df.loc[data18_df['target']==1], 'HLT_e28_lhtight_nod0_ringer_v13_ivarloose' , pidname='el_lhtight' , reset=True)
eff_data18_Zee.fill( data18_df.loc[data18_df['target']==1], 'HLT_e60_lhmedium_nod0_ringer_v13_L1EM24VHI', pidname='el_lhmedium', reset=True)
eff_data18_Zee.fill( data18_df.loc[data18_df['target']==1], 'HLT_e140_lhloose_nod0_ringer_v13'          , pidname='el_lhloose' , reset=True)

eff_data18_Zee.save()

eff_data18_JF17 = Efficiency( output_path+'/data18_JF17_efficiency_v13.root')

eff_data18_JF17.fill( data18_df.loc[data18_df['target']==0], 'HLT_e17_lhvloose_nod0_ringer_v13_L1EM15VHI', pidname='!el_lhvloose',reset=True)
eff_data18_JF17.fill( data18_df.loc[data18_df['target']==0], 'HLT_e28_lhtight_nod0_ringer_v13_ivarloose' , pidname='!el_lhvloose',reset=True)
eff_data18_JF17.fill( data18_df.loc[data18_df['target']==0], 'HLT_e60_lhmedium_nod0_ringer_v13_L1EM24VHI', pidname='!el_lhvloose',reset=True)
eff_data18_JF17.fill( data18_df.loc[data18_df['target']==0], 'HLT_e140_lhloose_nod0_ringer_v13'          , pidname='!el_lhvloose',reset=True)

eff_data18_JF17.save()

## Evaluate mc16 boosted:

mc16_df = pd.concat( (load_hdf(dpath_mc16), load_hdf(tpath_mc16)), axis=1)

emulate(mc16_df)

eff_mc16 = Efficiency( output_path+'/mc16_boosted_efficiency_v13.root')

eff_mc16.fill( mc16_df.loc[mc16_df['target']==1], 'HLT_e17_lhvloose_nod0_ringer_v13_L1EM15VHI', pidname='el_lhvloose', reset=True)
eff_mc16.fill( mc16_df.loc[mc16_df['target']==1], 'HLT_e28_lhtight_nod0_ringer_v13_ivarloose' , pidname='el_lhtight' , reset=True)
eff_mc16.fill( mc16_df.loc[mc16_df['target']==1], 'HLT_e60_lhmedium_nod0_ringer_v13_L1EM24VHI', pidname='el_lhmedium', reset=True)
eff_mc16.fill( mc16_df.loc[mc16_df['target']==1], 'HLT_e140_lhloose_nod0_ringer_v13'          , pidname='el_lhloose' , reset=True)

eff_mc16.save()

##
